initial checkin of C version of Universal Time Scale.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/3927



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@16587 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/unicode/pwin32.h
@@ -84,9 +84,8 @@
 #endif
 
 /* Define 64 bit limits */
-#define INT64_C(x) x
-#define UINT64_C(x) x
-
+#define INT64_C(x) ((int64_t)x)
+#define UINT64_C(x) ((uint64_t)x)
 
 /*===========================================================================*/
 /* Generic data types                                                        */
